refactor: improve popover, add optional blocking bg
