+++ src/scripts/ui/popover/index.ts
@@ -1,0 +1,4 @@
+import "./styles";
+
+export { confirm, type ConfirmOptions } from "./confirm";
+export { notify, type NotifyOptions } from "./notify";
